Make PETSc (non-)detection more robust
With PETSc installed using apt and an active Conda environment with the
conda-forge compiler but no petsc in it, the macro would detect the
system petsc, which would then fail to link because the c-f compiler
won't link to system libs. So this adds a check that we can link after
we've found the header, plus some cosmetic improvements.